refactor: add Car.kt and refactor Car class
