+++ src/main/kotlin/racingcar/model/Car.kt
@@ -1,0 +1,15 @@
+package racingcar.model
+
+import camp.nextstep.edu.missionutils.Randoms
+
+class Car(val name: String, var progress: Int = 0) {
+    fun move() {
+        if (Randoms.pickNumberInRange(0, 9) >= 4) {
+            progress++
+        }
+    }
+
+    fun getProgress(): String {
+        return "-".repeat(progress)
+    }
+}
